+++ package.json
@@ -50,6 +50,9 @@
     "expo-app-loading": "^1.0.1",
     "expo-font": "~8.4.0",
     "expo-splash-screen": "0.8.1",
+    "formik": "^2.2.9",
+    "i": "^0.3.7",
+    "npm": "^8.0.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-native": "https://github.com/expo/react-native/archive/sdk-40.0.1.tar.gz",
@@ -57,7 +60,8 @@
     "react-native-safe-area-context": "3.1.9",
     "react-native-screens": "~2.15.2",
     "react-native-svg": "^12.1.1",
-    "react-native-web": "0.17.1"
+    "react-native-web": "0.17.1",
+    "yup": "^0.32.11"
   },
   "devDependencies": {
     "@babel/core": "7.15.0",
