+++ src/types/initialValues.tsx
@@ -1,0 +1,8 @@
+interface InitialValues {
+  JarInformation: {
+    title: string | undefined
+    iconName: string | undefined
+  }
+}
+
+export default InitialValues
